fixed sign-up error, created friendship controller
