+++ app/controllers/posts_controller.rb
@@ -1,5 +1,6 @@
 class PostsController < ApplicationController
   before_action :authenticate_user!
+  include FriendshipHelper
 
   def index
     @post = Post.new
