+++ app/views/users/index.html.erb
@@ -1,5 +1,7 @@
 <section class="users-section">
   <h2>All Users</h2>
+  <p><%= @confirmed_friends %></p>
+  <p><%= @pending_friends %></p>
   <ul class="users-list">
     <% @users.each do |user| %>
       <li>
@@ -7,7 +9,7 @@
         <span class="profile-link">
           <%= link_to 'See Profile',  user_path(user), class: 'profile-link' %>
           <%unless (@confirmed_friends + @pending_friends).include?(user.id) || user.id == current_user.id%>
-            <%= button_to 'invite', invite_path, method: :post, params: { friend_id: user.id } %>
+            <%= button_to 'invite', friendships_path, method: :post, params: { friend_id: user.id } %>
           <% end %>
         </span>
       </li>
