+++ app/views/users/show.html.erb
@@ -1,7 +1,7 @@
 <section class="user-section">
   <h2><%= "Name: #{@user.name}" %></h2>
   <%unless (@confirmed_friends + @pending_friends).include?(@user.id) || @user.id == current_user.id%>
-    <%= button_to 'invite', invite_path, method: :post, params: { friend_id: @user.id } %>
+    <%= button_to 'invite', friendships_path, method: :post, params: { friend_id: @user.id } %>
   <% end %>
   <article class="timeline">
     <h3>Recent posts:</h3>
